third_party/WebKit/LayoutTests/imported/web-platform-tests/shadow-dom/Element-interface-attachShadow-expected.txt

This is a testharness.js-based test.
PASS Check the existence of Element.attachShadow 
PASS Nodes other than Element should not have attachShadow 
PASS Element.attachShadow must throw a TypeError if mode is not "open" or "closed" 
PASS Element.attachShadow must create an instance of ShadowRoot 
PASS Element.attachShadow must throw a InvalidStateError if the context object already hosts a shadow tree 
FAIL Element.attachShadow must throw a NotSupportedError for button, details, input, marquee, meter, progress, select, textarea, and keygen elements assert_throws: Calling attachShadow({mode: "open"}) on button element must throw function "function () {
            document.createElement(elementN..." did not throw
Harness: the test ran to completion.
